bump iterations 0 - 3 tests to match current spec
